+++ changelog.md
@@ -13,6 +13,8 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 ### Added
 
 - Area for notices below the "Come worship with us" section for short term updates
+- Notices can be stored in and loaded from database
+- Notices are filtered for the most current notice
 
 ### Changed